% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{provinces}
\alias{provinces}
\title{provinces}
\format{
A \code{list} object containing all the provinces.
}
\source{
Adapted from Natural Earth.
}
\usage{
provinces
}
\description{
Map of Canadian provinces.
}
\keyword{datasets}
