% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vaccination_coverage}
\alias{vaccination_coverage}
\title{vaccination_coverage}
\format{
A \code{data.frame} object containing all the provinces.
}
\source{
Adapted from the Government of Canada.
}
\usage{
vaccination_coverage
}
\description{
Cumulative percent of the population who have received at least 1 dose of a COVID-19 vaccine.
}
\keyword{datasets}
